requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flake8==7.3.0
black==25.1.0
